Support long marked offset in SlicedInputStream
Else, a reset might reset to an unintended offset in a big slice.

Relates ES-9639

+++ server/src/main/java/org/elasticsearch/index/snapshots/blobstore/SlicedInputStream.java
@@ -24,12 +24,12 @@ import java.io.InputStream;
 public abstract class SlicedInputStream extends InputStream {
     private int nextSlice = 0;
     private InputStream currentStream;
-    private int currentSliceOffset = 0;
+    private long currentSliceOffset = 0;
     private final int numSlices;
     private boolean closed = false;
     private boolean initialized = false;
     private int markedSlice = -1;
-    private int markedSliceOffset = -1;
+    private long markedSliceOffset = -1;
 
     /**
      * Creates a new SlicedInputStream
